I/GenericService with I/Repository was moved to their own libriary to avoid duplication
