+++ Clinic.API.sln
@@ -19,11 +19,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Notifications.API", "src\No
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Notifications.BLL", "src\Notofications\Notifications.BLL\Notifications.BLL.csproj", "{6F57AA9D-37C9-4457-9A6A-801E4031B0AF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Profiles.API", "src\Profiles\Profiles.API\Profiles.API.csproj", "{2B2C9515-4795-40BE-BAB2-AA7BAFBACCE0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Profiles.API", "src\Profiles\Profiles.API\Profiles.API.csproj", "{2B2C9515-4795-40BE-BAB2-AA7BAFBACCE0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Profiles.BLL", "src\Profiles\Profiles.BLL\Profiles.BLL.csproj", "{D9AD826F-2B8F-4B70-B01C-C4D5D60B806A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Profiles.BLL", "src\Profiles\Profiles.BLL\Profiles.BLL.csproj", "{D9AD826F-2B8F-4B70-B01C-C4D5D60B806A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Profiles.DAL", "src\Profiles\Profiles.DAL\Profiles.DAL.csproj", "{2D73B154-4B02-49B9-8EA0-350AA514BD80}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Profiles.DAL", "src\Profiles\Profiles.DAL\Profiles.DAL.csproj", "{2D73B154-4B02-49B9-8EA0-350AA514BD80}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StandartCRUD", "common\StandartCRUD\StandartCRUD.csproj", "{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -77,6 +79,10 @@ Global
 		{2D73B154-4B02-49B9-8EA0-350AA514BD80}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2D73B154-4B02-49B9-8EA0-350AA514BD80}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2D73B154-4B02-49B9-8EA0-350AA514BD80}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1CF5148-CD3F-4EE2-807E-A2F0EF19271C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
