+++ common/StandartCRUD/StandartDAL/Repository.cs
@@ -1,12 +1,13 @@
-﻿using Clinic.DAL.Interfaces;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using System.Linq.Expressions;
 
-namespace Clinic.DAL.Repositories
+namespace StandartCRUD.StandartDAL
 {
-    public class Repository<T>(ClinicDbContext dbContext) : IRepository<T> where T : class
+    public class Repository<T, TContext>(TContext dbContext) : 
+        IRepository<T> where T : class
+        where TContext : DbContext
     {
-        protected readonly ClinicDbContext _dbContext = dbContext;
+        protected readonly TContext _dbContext = dbContext;
 
         public async virtual Task<IEnumerable<T>> GetAllAsync(CancellationToken cancellationToken = default)
         {
