--- src/Clinic.BLL/Interfaces/IGenericService.cs
@@ -1,11 +1,0 @@
-﻿namespace Clinic.BLL.Interfaces
-{
-    public interface IGenericService<TEntity, TModel>
-    {
-        Task<TModel> GetByIdAsync(Guid id, CancellationToken cancellationToken);
-        Task<TModel> CreateAsync(TModel model, CancellationToken cancellationToken);
-        Task<TModel> UpdateAsync(TModel model, CancellationToken cancellationToken);
-        Task DeleteAsync(Guid id, CancellationToken cancellationToken);
-        Task<IEnumerable<TModel>> GetAllAsync(CancellationToken cancellationToken);
-    }
-}
