+++ src/Clinic.BLL/Interfaces/IOfficeService.cs
@@ -1,5 +1,6 @@
 ﻿using Clinic.BLL.Models;
 using Clinic.DAL.Entities;
+using StandartCRUD.StandartBLL;
 
 namespace Clinic.BLL.Interfaces
 {
