+++ src/Clinic.DAL/Clinic.DAL.csproj
@@ -16,4 +16,8 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\common\StandartCRUD\StandartCRUD.csproj" />
+  </ItemGroup>
+
 </Project>
