+++ src/Profiles/Profiles.BLL/Interrfaces/IPatientService.cs
@@ -1,5 +1,6 @@
 ﻿using Profiles.BLL.Models;
 using Profiles.DAL.Entities;
+using StandartCRUD.StandartBLL;
 
 namespace Profiles.BLL.Interrfaces
 {
