+++ src/Profiles/Profiles.BLL/Services/ReceptionistService.cs
@@ -3,6 +3,7 @@ using Profiles.BLL.Interrfaces;
 using Profiles.BLL.Models;
 using Profiles.DAL.Entities;
 using Profiles.DAL.Interfaces;
+using StandartCRUD.StandartBLL;
 
 namespace Profiles.BLL.Services
 {
