+++ src/Profiles/Profiles.DAL/Interfaces/IPatientRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Profiles.DAL.Entities;
+using StandartCRUD.StandartDAL;
 
 namespace Profiles.DAL.Interfaces
 {
