--- src/Profiles/Profiles.DAL/Interfaces/IRepository.cs
@@ -1,11 +1,0 @@
-﻿namespace Profiles.DAL.Interfaces
-{
-    public interface IRepository<T> where T : class
-    {
-        Task<T?> GetByIdAsync(Guid id, CancellationToken cancellationToken);
-        Task<IEnumerable<T>> GetAllAsync(CancellationToken cancellationToken);
-        Task<T> AddAsync(T entity, CancellationToken cancellationToken);
-        Task<T> UpdateAsync(T entity, CancellationToken cancellationToken);
-        Task DeleteAsync(T entity, CancellationToken cancellationToken);
-    }
-}
