+++ src/Profiles/Profiles.DAL/Profiles.DAL.csproj
@@ -20,4 +20,8 @@
     <Folder Include="Migrations\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\common\StandartCRUD\StandartCRUD.csproj" />
+  </ItemGroup>
+
 </Project>
