+++ src/Profiles/Profiles.DAL/Repositories/DoctorRepository.cs
@@ -1,11 +1,12 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Profiles.DAL.Entities;
 using Profiles.DAL.Interfaces;
+using StandartCRUD.StandartDAL;
 
 namespace Profiles.DAL.Repositories
 {
     public class DoctorRepository(ProfilesDbContext dbContext) : 
-        Repository<DoctorEntity>(dbContext), 
+        Repository<DoctorEntity, ProfilesDbContext>(dbContext), 
         IDoctorRepository
     {
         public async Task<IEnumerable<DoctorEntity>> GetAllAsync(string? name, DoctorStatus? status, CancellationToken cancellationToken)
