--- src/Profiles/Profiles.DAL/Repositories/Repository.cs
@@ -1,48 +1,0 @@
-﻿using Profiles.DAL.Interfaces;
-using Microsoft.EntityFrameworkCore;
-using System.Linq.Expressions;
-
-namespace Profiles.DAL.Repositories
-{
-    public class Repository<T>(ProfilesDbContext dbContext) : IRepository<T> where T : class
-    {
-        protected readonly ProfilesDbContext _dbContext = dbContext;
-
-        public async virtual Task<IEnumerable<T>> GetAllAsync(CancellationToken cancellationToken = default)
-        {
-            return await _dbContext.Set<T>().ToListAsync(cancellationToken);
-        }
-
-        public async virtual Task<T?> GetByIdAsync(Guid id, CancellationToken cancellationToken)
-        {
-            return await _dbContext.Set<T>().FindAsync([id, cancellationToken], cancellationToken: cancellationToken);
-        }
-
-        public async virtual Task<T> AddAsync(T entity, CancellationToken cancellationToken)
-        {
-            await _dbContext.Set<T>().AddAsync(entity, cancellationToken);
-            await _dbContext.SaveChangesAsync(cancellationToken);
-
-            return entity;
-        }
-
-        public async virtual Task<T> UpdateAsync(T entity, CancellationToken cancellationToken)
-        {
-            _dbContext.Set<T>().Update(entity);
-            await _dbContext.SaveChangesAsync(cancellationToken);
-
-            return entity;
-        }
-
-        public async virtual Task DeleteAsync(T entity, CancellationToken cancellationToken)
-        {
-            _dbContext.Set<T>().Remove(entity);
-            await _dbContext.SaveChangesAsync(cancellationToken);
-        }
-
-        public async virtual Task<IEnumerable<T>> GetByPredicateAsync(Expression<Func<T, bool>> predicate, CancellationToken cancellationToken)
-        {
-            return await _dbContext.Set<T>().Where(predicate).ToListAsync(cancellationToken);
-        }
-    }
-}
